Rename toggle_command_bar_action files names to set_mode_action
