+++ include/fullscore/actions/set_mode_action.h
